+++ lib/sm_input_field.dart
@@ -1,5 +1,5 @@
 library sm_input_field;
 
-export 'fields/input_field.dart';
-export 'fields/search_field.dart';
-export 'models/field_info.dart';
+export 'fields/input_field.dart' show InputField;
+export 'fields/search_field.dart' show SearchField;
+export 'models/field_info.dart' show FieldInfo;