+++ lib/utils/constants.dart
@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
 
+// This file contains the different constants used in this package
+
+// colors
 const kColorBlue = Color(0xff43BFC7), kColorPurple = Color(0xff40356F);
 
+// text styles
 const kTextStyleMain = TextStyle(fontSize: 18);