+++ CHANGELOG.md
@@ -1,3 +1,19 @@
 ## 1.0.0
 
-* TODO: Describe initial release.
+Initial Release
+
+## 1.0.1
+
+<h2>Changed:</h2>
+<ul>
+    <li>Updated README</li>
+    <li>Updated package with doc comments</li>
+</ul>
+<br>
+<h2>Added:</h2>
+<ul>
+    <li>Added example project</li>
+</ul>
+
+
+